requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==5.1.5
django-allauth==65.3.1
django-crispy-forms==2.3
django-environ==0.4.5
django-filter==24.3
djangorestframework==3.15.2
inflection==0.5.1
packaging==24.2
psycopg2-binary==2.9.10
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
gunicorn==23.0.0
django-extensions==3.2.3